+++ .env.example
@@ -1,3 +1,6 @@
+# The path from which to fetch the proposals manifest. Empty defaults to the MinaFoundation/mina-on-chain-voting GitHub.
+PROPOSALS_URL=""
+
 BUCKET_NAME="673156464838-mina-staking-ledgers"
 
 # [REQUIRED] - the connection URL for the archive database.
