--- server/rust-toolchain
@@ -1,1 +1,0 @@
-stable
