+++ server/src/handlers.rs
@@ -1,9 +1,9 @@
 mod get_mina_proposal;
 mod get_mina_proposal_result;
 mod get_mina_proposals;
-mod info;
+mod get_info;
 
 pub use get_mina_proposal::*;
 pub use get_mina_proposal_result::*;
 pub use get_mina_proposals::*;
-pub use info::*;
+pub use get_info::*;
