+++ server/src/handlers/get_info.rs
